Squash the whole trait inside the wrapping connection
