--- src/Driver/PDO/MySQL/Driver.php
@@ -1,64 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\Driver\PDO\MySQL;
-
-use Doctrine\DBAL\Driver\AbstractMySQLDriver;
-use Doctrine\DBAL\Driver\PDO;
-use Facile\DoctrineMySQLComeBack\Doctrine\DBAL\Driver\ServerGoneAwayExceptionsAwareInterface;
-use Facile\DoctrineMySQLComeBack\Doctrine\DBAL\Driver\ServerGoneAwayExceptionsAwareTrait;
-
-/**
- * Class Driver.
- */
-class Driver extends AbstractMySQLDriver implements ServerGoneAwayExceptionsAwareInterface
-{
-    use ServerGoneAwayExceptionsAwareTrait;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function connect(array $params, $username = null, $password = null, array $driverOptions = [])
-    {
-        return new PDO\Connection(
-            $this->constructPdoDsn($params),
-            $username ?? $params['user'] ?? '',
-            $password ?? $params['password'] ?? '',
-            $driverOptions
-        );
-    }
-
-    /**
-     * Constructs the MySql PDO DSN.
-     *
-     * @param mixed[] $params
-     *
-     * @return string the DSN
-     */
-    private function constructPdoDsn(array $params): string
-    {
-        $dsn = 'mysql:';
-        if (isset($params['host']) && $params['host'] !== '') {
-            $dsn .= 'host=' . $params['host'] . ';';
-        }
-
-        if (isset($params['port'])) {
-            $dsn .= 'port=' . $params['port'] . ';';
-        }
-
-        if (isset($params['dbname'])) {
-            $dsn .= 'dbname=' . $params['dbname'] . ';';
-        }
-
-        if (isset($params['unix_socket'])) {
-            $dsn .= 'unix_socket=' . $params['unix_socket'] . ';';
-        }
-
-        if (isset($params['charset'])) {
-            $dsn .= 'charset=' . $params['charset'] . ';';
-        }
-
-        return $dsn;
-    }
-}
